// cgocore-ui/src/app/features/models/shipments/views/basic/AirWaybillBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirWaybillBasicView {

id: number;
selected: boolean;
canDelete: boolean;
airWaybillNumber: string;
forwarderName: string;
originIataCode: string;
destinationIataCode: string;
goddsDescription: string;
hsCode: string;
numberOfPieces: number;
grossWeight: number;
chargeableWeight: number;
grossWeightUnitOfMeasureCode: string;
grossVolume: number;
grossVolumeUnitOfMeasureCode: string;

/* Specific Code Start */
/* Specific Code End */
}