// cgocore-ui/src/app/features/models/shipments/filters/AirWaybillFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirWaybillFilter {

airWaybillNumber: string;
forwarderName: string;
originIataCode: string;
destinationIataCode: string;
goddsDescription: string;
hsCode: string;

/* Specific Code Start */
/* Specific Code End */
}