// cgocore-ui/src/app/features/models/shipments/sortings/AirWaybillSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirWaybillSorting {

airWaybillNumberOrderType: string;
forwarderNameOrderType: string;
originIataCodeOrderType: string;
destinationIataCodeOrderType: string;
goddsDescriptionOrderType: string;
hsCodeOrderType: string;
numberOfPiecesOrderType: string;
grossWeightOrderType: string;
chargeableWeightOrderType: string;
grossWeightUnitOfMeasureCodeOrderType: string;
grossVolumeOrderType: string;
grossVolumeUnitOfMeasureCodeOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}